requirements.txt
neuromorphic_drivers==0.10.1
pyserial==3.5
